Evaluate multiple expressions created with symbol substitution (raw version).

Loop over a grid of values, create an expression object from each one, and then evaluate that expression. Helps with general metaprogramming. Unlike tar_sub(), which quotes the expr argument, tar_sub_raw() assumes expr is an expression object.

tar_eval_raw(expr, values, envir = parent.frame())

Arguments

expr
Expression object with the starting expression. Values are iteratively substituted in place of symbols in expr to create each new expression, and then each expression is evaluated.

values
List of values to substitute into expr to create the expressions. All elements of values must have the same length.

envir
Environment in which to evaluate the new expressions.

Value

A list of return values from evaluating the expression objects. Often, these values are target objects. Target objects represent skippable steps of the analysis pipeline as described at https://books.ropensci.org/targets/. Please see the design specification at https://books.ropensci.org/targets-design/ to learn about the structure and composition of target objects.

Examples

# tar_map() is incompatible with tar_render() because the latter
# operates on preexisting tar_target() objects. By contrast,
# tar_eval_raw() and tar_sub_raw() iterate over code farther upstream.
values <- list(
  name = rlang::syms(c("name1", "name2")),
  file = c("file1.Rmd", "file2.Rmd")
)
tar_sub_raw(quote(list(name, file)), values = values)

#> [[1]]
#> expression(list(name1, "file1.Rmd"))
#> 
#> [[2]]
#> expression(list(name2, "file2.Rmd"))
#> 

tar_sub_raw(quote(tar_render(name, file)), values = values)

#> [[1]]
#> expression(tar_render(name1, "file1.Rmd"))
#> 
#> [[2]]
#> expression(tar_render(name2, "file2.Rmd"))
#> 

path <- tempfile()
file.create(path)

#> [1] TRUE

str(tar_eval_raw(quote(tar_render(name, path)), values = values))

#> List of 2
#>  $ :Classes 'tar_stem', 'tar_builder', 'tar_target', 'environment' <environment: 0x55ca1cb7a648> 
#>  $ :Classes 'tar_stem', 'tar_builder', 'tar_target', 'environment' <environment: 0x55ca1ccf0648> 

# So in your _targets.R file, you can define a pipeline like as below.
# Just make sure to set a unique name for each target
# (which tar_map() does automatically).
values <- list(
  name = rlang::syms(c("name1", "name2")),
  file = c(path, path)
)
list(
  tar_eval_raw(quote(tar_render(name, file)), values = values)
)

#> [[1]]
#> [[1]][[1]]
#> <stem target> 
#>   name: name1 
#>   command:
#>     tarchetypes::tar_render_run(path = "/tmp/Rtmp1rRQ79/file2ca6ee0e0e9", 
#>         args = list(input = "/tmp/Rtmp1rRQ79/file2ca6ee0e0e9", knit_root_dir = getwd(), 
#>             quiet = TRUE), deps = list()) 
#>   format: file 
#>   iteration method: vector 
#>   error mode: stop 
#>   memory mode: persistent 
#>   storage mode: main 
#>   retrieval mode: main 
#>   deploy to: main 
#>   resources:
#>     list() 
#>   cue:
#>     mode: thorough
#>     command: TRUE
#>     depend: TRUE
#>     format: TRUE
#>     iteration: TRUE
#>     file: TRUE 
#>   packages:
#>     tarchetypes
#>     stats
#>     graphics
#>     grDevices
#>     utils
#>     datasets
#>     methods
#>     base 
#>   library:
#>     NULL
#> [[1]][[2]]
#> <stem target> 
#>   name: name2 
#>   command:
#>     tarchetypes::tar_render_run(path = "/tmp/Rtmp1rRQ79/file2ca6ee0e0e9", 
#>         args = list(input = "/tmp/Rtmp1rRQ79/file2ca6ee0e0e9", knit_root_dir = getwd(), 
#>             quiet = TRUE), deps = list()) 
#>   format: file 
#>   iteration method: vector 
#>   error mode: stop 
#>   memory mode: persistent 
#>   storage mode: main 
#>   retrieval mode: main 
#>   deploy to: main 
#>   resources:
#>     list() 
#>   cue:
#>     mode: thorough
#>     command: TRUE
#>     depend: TRUE
#>     format: TRUE
#>     iteration: TRUE
#>     file: TRUE 
#>   packages:
#>     tarchetypes
#>     stats
#>     graphics
#>     grDevices
#>     utils
#>     datasets
#>     methods
#>     base 
#>   library:
#>     NULL
#>