Easy branching over dynamic files.

Shorthand for a pattern that correctly branches over files.

tar_files(
  name,
  command,
  tidy_eval = targets::tar_option_get("tidy_eval"),
  packages = targets::tar_option_get("packages"),
  library = targets::tar_option_get("library"),
  format = targets::tar_option_get("format"),
  iteration = targets::tar_option_get("iteration"),
  error = targets::tar_option_get("error"),
  memory = targets::tar_option_get("memory"),
  template = targets::tar_option_get("template"),
  deployment = targets::tar_option_get("deployment"),
  priority = 0,
  resources = targets::tar_option_get("resources"),
  storage = targets::tar_option_get("storage"),
  retrieval = targets::tar_option_get("retrieval"),
  cue = targets::tar_option_get("cue")
)

Arguments

name
Character of length 1, name of the target.

command
Similar to the command argument of tar_target() except the object must already be an expression instead of informally quoted code. base::expression() and base::quote() can produce such objects.

tidy_eval
Whether to invoke tidy evaluation (e.g. the !! operator from rlang) as soon as the target is defined (before tar_make()). Applies to the command argument.

packages
Character vector of packages to load right before the target builds. Use tar_option_set() to set packages globally for all subsequent targets you define.

library
Character vector of library paths to try when loading packages.

format
Optional storage format for the target's return value. With the exception of format = "file", each target gets a file in _targets/objects, and each format is a different way to save and load this file. Possible formats:

"file": A dynamic file. To use this format, the target needs to manually identify or save some data and return a character vector of paths to the data. Those paths must point to files or directories, and they must not contain characters | or *. Then, targets automatically checks those files and cues the appropriate build decisions if those files are out of date.

"rds": Default, uses saveRDS() and readRDS(). Should work for most objects, but slow.

"qs": Uses qs::qsave() and qs::qread(). Should work for most objects, much faster than "rds".

"fst": Uses fst::write_fst() and fst::read_fst(). Much faster than "rds", but the value must be a data frame.

"fst_dt": Same as "fst", but the value is a data.table.

"fst_tbl": Same as "fst", but the value is a tibble.

"keras": Uses keras::save_model_hdf5() and keras::load_model_hdf5(). The value must be a Keras model.

iteration
Character of length 1, name of the iteration mode of the target. Choices:

"vector": branching happens with vectors::vec_slice() and aggregation happens with vctrs::vec_c().

"list", branching happens with [[]] and aggregation happens with list().

"group": dplyr::group_by()-like functionality to branch over subsets of a data frame. The target's return value must be a data frame with a special tar_group column of consecutive integers from 1 through the number of groups. Each integer designates a group, and a branch is created for each collection of rows in a group. See the tar_group() function to see how you can create the special tar_group column with dplyr::group_by().

error
Character of length 1, what to do if the target runs into an error. If "stop", the whole pipeline stops and throws an error. If "continue", the error is recorded, but the pipeline keeps going.

memory
Character of length 1, memory strategy. If "persistent", the target stays in memory until the end of the pipeline. If "transient", the target gets unloaded after every new target completes. Either way, the target gets automatically loaded into memory whenever another target needs the value.

template
Relevant to tar_make_clustermq() only. Named list of values to fill in the clustermq template file. Unsupported for now. May be supported in the future if clustermq ever supports heterogeneous workers with varying resource requirements. In the meantime, use the template argument of tar_make_clustermq().

deployment
Character of length 1, only relevant to tar_make_clustermq() and tar_make_future(). If "remote", the target builds on a remote parallel worker. If "local", the target builds on the host machine / process managing the pipeline.

priority
Numeric of length 1 between 0 and 1. Controls which targets get deployed first when multiple competing targets are ready simultaneously. Targets with priorities closer to 1 get built earlier.

resources
Relevant to tar_make_future() only. A named list of resources passed to future::future() when defining a new worker.

storage
Character of length 1, only relevant to tar_make_clustermq() and tar_make_future(). If "local", the target's return value is sent back to the host machine and saved locally. If "remote", the remote worker saves the value.

retrieval
Character of length 1, only relevant to tar_make_clustermq() and tar_make_future(). If "local", the target's dependencies are loaded on the host machine and sent to the remote worker before the target builds. If "remote", the remote worker loads the targets dependencies.

cue
An optional object from tar_cue() to customize the rules that decide whether the target is up to date.

Value

A list of two targets, one upstream and one downstream. The upstream one does some work and returns some file paths, and the downstream target is a pattern that applies format = "file".

Details

tar_files() creates a pair of targets, one upstream and one downstream. The upstream target does some work and returns some file paths, and the downstream target is a pattern that applies format = "file". This is the correct way to dynamically iterate over file targets. It makes sure any downstream patterns only rerun some of their branches if the files change. For more information, visit https://github.com/wlandau/targets/issues/136 and https://github.com/ropensci/drake/issues/1302.

Examples

if (FALSE) {
# Without loss of generality,
tar_files(your_target, c("a.txt", "b.txt"))
# is equivalent to:
list(
  tar_target(your_target_files, c("a.txt", "b.txt")),
  tar_target(
    your_target,
    your_target_files,
    pattern = map(your_target_files),
    format = "file",
    deployment = "local",
    storage = "local",
    retrieval = "local"
  )
)
# Try it out.
file.create(c("a.txt", "b.txt"))
targets::tar_dir({
targets::tar_script({
  tar_pipeline(
    tarchetypes::tar_files(x, c("a.txt", "b.txt"))
  )
})
targets::tar_make()
# Should be up to date now.
targets::tar_make()
# If we change one file, `tar_make()` will only rerun one branch.
writeLines("a", "a.txt")
tar_make()
})
}