Quarto file detection

Detect the important files in a Quarto project.

Usage

tar_quarto_files(path = ".", profile = NULL)

Arguments

path

Character of length 1, either the file path to a Quarto source document or the directory path to a Quarto project. Defaults to the Quarto project in the current working directory.

profile

Character of length 1, Quarto profile. If NULL, the default profile will be used. Requires Quarto version 1.2 or higher. See https://quarto.org/docs/projects/profiles.html for details.

Value

A named list of important file paths in a Quarto project or document:

sources: source files with tar_load()/tar_read() target dependencies in R code chunks.

output: output files that will be generated during quarto::quarto_render().

input: pre-existing files required to render the project or document, such as _quarto.yml.

Details

This function is just a thin wrapper that interprets the output of quarto::quarto_inspect() and returns what tarchetypes needs to know about the current Quarto project or document.

See also

Examples

lines <- c(
  "---",
  "title: source file",
  "---",
  "Assume these lines are in report.qmd.",
  "```{r}",
  "1 + 1",
  "```"
)
path <- tempfile(fileext = ".qmd")
writeLines(lines, path)
# If Quarto is installed, run:
# tar_quarto_files(path)