Batched replication with dynamic branching (raw version).

Shorthand for a pattern that replicates a command using batches. Batches reduce the number of targets and thus reduce overhead.

tar_rep_raw(
  name,
  command,
  batches = 1,
  reps = 1,
  tidy_eval = targets::tar_option_get("tidy_eval"),
  packages = targets::tar_option_get("packages"),
  library = targets::tar_option_get("library"),
  format = targets::tar_option_get("format"),
  iteration = targets::tar_option_get("iteration"),
  error = targets::tar_option_get("error"),
  memory = targets::tar_option_get("memory"),
  garbage_collection = targets::tar_option_get("garbage_collection"),
  deployment = targets::tar_option_get("deployment"),
  priority = targets::tar_option_get("priority"),
  resources = targets::tar_option_get("resources"),
  storage = targets::tar_option_get("storage"),
  retrieval = targets::tar_option_get("retrieval"),
  cue = targets::tar_option_get("cue")
)

Arguments

name
Character of length 1, name of the target.

command
Expression object with code to run multiple times. Must return a list or data frame when evaluated.

batches
Number of batches. This is also the number of dynamic branches created during tar_make().

reps
Number of replications in each batch. The total number of replications is batches * reps.

tidy_eval
Whether to invoke tidy evaluation (e.g. the !! operator from rlang) as soon as the target is defined (before tar_make()). Applies to the command argument.

packages
Character vector of packages to load right before the target builds. Use tar_option_set() to set packages globally for all subsequent targets you define.

library
Character vector of library paths to try when loading packages.

format
Optional storage format for the target's return value. With the exception of format = "file", each target gets a file in _targets/objects, and each format is a different way to save and load this file. Possible formats:

"rds": Default, uses saveRDS() and readRDS(). Should work for most objects, but slow.

"qs": Uses qs::qsave() and qs::qread(). Should work for most objects, much faster than "rds". Optionally set the preset for qsave() through the resources argument, e.g. tar_target(..., resources = list(preset = "archive")).

"fst": Uses fst::write_fst() and fst::read_fst(). Much faster than "rds", but the value must be a data frame. Optionally set the compression level for fst::write_fst() through the resources argument, e.g. tar_target(..., resources = list(compress = 100)).

"fst_dt": Same as "fst", but the value is a data.table. Optionally set the compression level the same way as for "fst".

"fst_tbl": Same as "fst", but the value is a tibble. Optionally set the compression level the same way as for "fst".

"keras": Uses keras::save_model_hdf5() and keras::load_model_hdf5(). The value must be a Keras model.

"torch": Uses torch::torch_save() and torch::torch_load(). The value must be an object from the torch package such as a tensor or neural network module.

"file": A dynamic file. To use this format, the target needs to manually identify or save some data and return a character vector of paths to the data. Then, targets automatically checks those files and cues the appropriate build decisions if those files are out of date. Those paths must point to files or directories, and they must not contain characters | or *. All the files and directories you return must actually exist, or else targets will throw an error. (And if storage is "worker", targets will first stall out trying to wait for the file to arrive over a network file system.)

"url": A dynamic input URL. It works like format = "file" except the return value of the target is a URL that already exists and serves as input data for downstream targets. Optionally supply a custom curl handle through the resources argument, e.g. tar_target(..., resources = list(handle = curl::new_handle())). The data file at the URL needs to have an ETag or a Last-Modified time stamp, or else the target will throw an error because it cannot track the data. Also, use extreme caution when trying to use format = "url" to track uploads. You must be absolutely certain the ETag and Last-Modified time stamp are fully updated and available by the time the target's command finishes running. targets makes no attempt to wait for the web server.

"aws_rds", "aws_qs", "aws_fst", "aws_fst_dt", "aws_fst_tbl", "aws_keras": AWS-powered versions of the respective formats "rds", "qs", etc. The only difference is that the data file is uploaded to the AWS S3 bucket you supply to resources. See the cloud computing chapter of the manual for details.

"aws_file": arbitrary dynamic files on AWS S3. The target should return a path to a temporary local file, then targets will automatically upload this file to an S3 bucket and track it for you. Unlike format = "file", format = "aws_file" can only handle one single file, and that file must not be a directory. tar_read() and downstream targets download the file to _targets/scratch/ locally and return the path. _targets/scratch/ gets deleted at the end of tar_make(). Requires the same resources and other configuration details as the other AWS-powered formats. See the cloud computing chapter of the manual for details.

iteration
Character of length 1, name of the iteration mode of the target. Choices:

"vector": branching happens with vctrs::vec_slice() and aggregation happens with vctrs::vec_c().

"list", branching happens with [[]] and aggregation happens with list().

"group": dplyr::group_by()-like functionality to branch over subsets of a data frame. The target's return value must be a data frame with a special tar_group column of consecutive integers from 1 through the number of groups. Each integer designates a group, and a branch is created for each collection of rows in a group. See the tar_group() function to see how you can create the special tar_group column with dplyr::group_by().

error
Character of length 1, what to do if the target runs into an error. If "stop", the whole pipeline stops and throws an error. If "continue", the error is recorded, but the pipeline keeps going.

memory
Character of length 1, memory strategy. If "persistent", the target stays in memory until the end of the pipeline (unless storage is "worker", in which case targets unloads the value from memory right after storing it in order to avoid sending copious data over a network). If "transient", the target gets unloaded after every new target completes. Either way, the target gets automatically loaded into memory whenever another target needs the value. For cloud-based dynamic files such as format = "aws_file", this memory policy applies to temporary local copies of the file in _targets/scratch/": "persistent" means they remain until the end of the pipeline, and "transient" means they get deleted from the file system as soon as possible. The former conserves bandwidth, and the latter conserves local storage.

garbage_collection
Logical, whether to run base::gc() just before the target runs.

deployment
Character of length 1, only relevant to tar_make_clustermq() and tar_make_future(). If "worker", the target builds on a parallel worker. If "main", the target builds on the host machine / process managing the pipeline.

priority
Numeric of length 1 between 0 and 1. Controls which targets get deployed first when multiple competing targets are ready simultaneously. Targets with priorities closer to 1 get built earlier. Only applies to tar_make_future() and tar_make_clustermq() (not tar_make()). tar_make_future() with no extra settings is a drop-in replacement for tar_make() in this case.

resources
A named list of computing resources. Uses:

Template file wildcards for future::future() in tar_make_future().

Template file wildcards clustermq::workers() in tar_make_clustermq().

Custom target-level future::plan(), e.g. resources = list(plan = future.callr::callr).

Custom curl handle if format = "url", e.g. resources = list(handle = curl::new_handle()).

Custom preset for qs::qsave() if format = "qs", e.g. resources = list(handle = "archive").

Custom compression level for fst::write_fst() if format is "fst", "fst_dt", or "fst_tbl", e.g. resources = list(compress = 100).

AWS bucket and prefix for the "aws_" formats, e.g. resources = list(bucket = "your-bucket", prefix = "folder/name"). bucket is required for AWS formats. See the cloud computing chapter of the manual for details.

storage
Character of length 1, only relevant to tar_make_clustermq() and tar_make_future(). If "main", the target's return value is sent back to the host machine and saved locally. If "worker", the worker saves the value.

retrieval
Character of length 1, only relevant to tar_make_clustermq() and tar_make_future(). If "main", the target's dependencies are loaded on the host machine and sent to the worker before the target builds. If "worker", the worker loads the targets dependencies.

cue
An optional object from tar_cue() to customize the rules that decide whether the target is up to date.

Value

A list of two targets, one upstream and one downstream. The upstream one does some work and returns some file paths, and the downstream target is a pattern that applies format = "file".

Details

tar_rep() and tar_rep_raw each create two targets: an upstream local stem with an integer vector of batch ids, and a downstream pattern that maps over the batch ids. (Thus, each batch is a branch.) Each batch/branch replicates the command a certain number of times.

Both batches and reps within each batch are aggregated according to the method you specify in the iteration argument. If "list", reps and batches are aggregated with list(). If "vector", then vctrs::vec_c(). If "group", then vctrs::vec_rbind().

Examples

if (identical(Sys.getenv("TARCHETYPES_LONG_EXAMPLES"), "true")) {
targets::tar_dir({
targets::tar_script({
  targets::tar_pipeline(
    tarchetypes::tar_rep_raw(
      "x",
      expression(data.frame(x = sample.int(1e4, 2))),
      batches = 2,
      reps = 3
    )
  )
})
targets::tar_make(callr_function = NULL)
targets::tar_read(x)
})
}