Evaluate multiple expressions created with symbol substitution (raw version).

Loop over a grid of values, create an expression object from each one, and then evaluate that expression. Helps with general metaprogramming. Unlike tar_sub(), which quotes the expr argument, tar_sub_raw() assumes expr is an expression object.

tar_eval_raw(expr, values, envir = parent.frame())

Arguments

expr
Expression object with the starting expression. Values are iteratively substituted in place of symbols in expr to create each new expression, and then each expression is evaluated.

values
List of values to substitute into expr to create the expressions. All elements of values must have the same length.

envir
Environment in which to evaluate the new expressions.

Value

A list of return values from evaluating the expression objects.

Examples

# tar_map() is incompatible with tar_render() because the latter
# operates on preexisting tar_target() objects. By contrast,
# tar_eval_raw() and tar_sub_raw() iterate over code farther upstream.
values <- list(
  name = rlang::syms(c("name1", "name2")),
  file = c("file1.Rmd", "file2.Rmd")
)
tar_sub_raw(quote(tar_render(name, file)), values = values)

#> [[1]]
#> expression(tar_render(name1, "file1.Rmd"))
#> 
#> [[2]]
#> expression(tar_render(name2, "file2.Rmd"))
#> 

targets::tar_dir({
file.create(c("file1.Rmd", "file2.Rmd"))
str(tar_eval_raw(quote(tar_render(name, file)), values = values))
# So in your _targets.R file, you can define a pipeline like as below.
# Just make sure to set a unique name for each target
# (which tar_map() does automatically).
values <- list(
  name = rlang::syms(c("name1", "name2")),
  file = c("file1.Rmd", "file2.Rmd")
)
targets::tar_pipeline(
  tar_eval_raw(quote(tar_render(name, file)), values = values)
)
})

#> List of 2
#>  $ :Classes 'tar_stem', 'tar_builder', 'tar_target', 'environment' <environment: 0x7ffc2171ab50> 
#>  $ :Classes 'tar_stem', 'tar_builder', 'tar_target', 'environment' <environment: 0x7ffc240b2a28> 

#> <pipeline with 2 targets>