Static branching.

Define collections of new targets.

tar_map(values, ..., names = tidyselect::everything(), unlist = TRUE)

Arguments

values
Named list or data frame with values to iterate over. The names are the names of symbols in the commands and pattern statements, and the elements are values that get substituted in place of those symbols. Elements of the values list should be small objects that can easily deparse to names, such as characters, integers, and symbols. To create a list of symbols as a column of values, use rlang::syms().

...
One or more target objects or list of target objects. Lists can be arbitrarily nested, as in targets::tar_pipeline().

names
Subset of names(values) used to generate the suffixes in the names of the new targets. You can supply symbols, a character vector, or tidyselect helpers like starts_with().

unlist
Logical, whether to flatten the returned list of targets. If unlist = FALSE, the list is nested and sub-lists are grouped by the original input targets.

Value

A list of new target objects. If unlist is FALSE, the list is nested and sub-lists are grouped by the original input targets.

Details

tar_map() creates collections of new targets by iterating over a list of arguments and substituting symbols into commands and pattern statements.

Examples

targets::tar_dir({
targets::tar_script({
  targets::tar_pipeline(
    tarchetypes::tar_map(
      list(a = c(12, 34), b = c(45, 78)),
      targets::tar_target(x, a + b),
      targets::tar_target(y, x + a, pattern = map(x))
    )
  )
})
targets::tar_manifest(callr_function = NULL)
})

#> # A tibble: 4 x 4
#>   name    command      type  dimensions
#>   <chr>   <chr>        <chr> <list>    
#> 1 y_12_45 x_12_45 + 12 map   <chr [1]> 
#> 2 x_12_45 12 + 45      stem  <chr [0]> 
#> 3 y_34_78 x_34_78 + 34 map   <chr [1]> 
#> 4 x_34_78 34 + 78      stem  <chr [0]>